refactor: 이미지 너비,높이 객체 분리
